refactor: split vc meeting shortcuts
